Allow leagues to be tracked across seasons
This commit saves the initial work on redesigning the core data model and thus the rest of the application for this matter.
I am currently abandoning this idea, because it would be too much of an effort to harmonize historical data.
In concrete terms one example is that league names and abbreviations have changed over time.